Consolidate fixtures and specialize test modules
